[FIX/#38] 학과 불러오기 api 로직 수정

+++ data/src/main/java/com/example/data/datasource/OnboardingDataSource.kt
@@ -21,7 +21,7 @@ interface OnboardingDataSource {
     suspend fun getDepartmentNameData(
         school: String,
         search: String,
-        page: Long,
+        page: Int,
     ): BaseResponse<ResponseDepartmentDto>
 
     suspend fun getIdValidData(
